api/requirements.txt
fastapi
sqlalchemy
psycopg2-binary
jwt
python-jose[cryptography]
passlib[bcrypt]
python-multipart